Moved Specifics and Flyweight events from CrossComponentGraphItertor to AbstractGraphIterator.
