+++ jgrapht-core/src/main/java/org/jgrapht/traverse/AbstractGraphIterator.java
@@ -25,7 +25,7 @@
  * (C) Copyright 2003-2008, by Barak Naveh and Contributors.
  *
  * Original Author:  Barak Naveh
- * Contributor(s):   Christian Hammer
+ * Contributor(s):   Christian Hammer, Assaf Mizrachi
  *
  * $Id$
  *
@@ -34,12 +34,15 @@
  * 24-Jul-2003 : Initial revision (BN);
  * 11-Aug-2003 : Adaptation to new event model (BN);
  * 04-May-2004 : Made generic (CH)
+ * 15-Jul-2016 : Moved Specifics and Flyweight events from CrossComponentGraphItertor
  *
  */
 package org.jgrapht.traverse;
 
 import java.util.*;
 
+import org.jgrapht.DirectedGraph;
+import org.jgrapht.Graph;
 import org.jgrapht.event.*;
 
 
@@ -62,6 +65,11 @@ public abstract class AbstractGraphIterator<V, E>
     // so that subclasses can use it as a fast check to see if
     // event firing calls can be skipped.
     protected int nListeners = 0;
+    // TODO: support ConcurrentModificationException if graph modified
+    // during iteration.
+    protected FlyweightEdgeEvent<V, E> reusableEdgeEvent;
+    protected FlyweightVertexEvent<V> reusableVertexEvent;
+    protected Specifics<V, E> specifics;
 
     /**
      * Sets the cross component traversal flag - indicates whether to traverse
@@ -203,6 +211,159 @@ public abstract class AbstractGraphIterator<V, E>
         for (int i = 0; i < nListeners; i++) {
             TraversalListener<V, E> l = traversalListeners.get(i);
             l.vertexFinished(e);
+        }
+    }
+    
+    /**
+     * A reusable edge event.
+     *
+     * @author Barak Naveh
+     * @since Aug 11, 2003
+     */
+    static class FlyweightEdgeEvent<VV, localE>
+        extends EdgeTraversalEvent<localE>
+    {
+        private static final long serialVersionUID = 4051327833765000755L;
+
+        /**
+         * @see EdgeTraversalEvent
+         */
+        public FlyweightEdgeEvent(Object eventSource, localE edge)
+        {
+            super(eventSource, edge);
+        }
+
+        /**
+         * Sets the edge of this event.
+         *
+         * @param edge the edge to be set.
+         */
+        protected void setEdge(localE edge)
+        {
+            this.edge = edge;
+        }
+    }
+
+    /**
+     * A reusable vertex event.
+     *
+     * @author Barak Naveh
+     * @since Aug 11, 2003
+     */
+    static class FlyweightVertexEvent<VV>
+        extends VertexTraversalEvent<VV>
+    {
+        private static final long serialVersionUID = 3834024753848399924L;
+
+        /**
+         * @see VertexTraversalEvent#VertexTraversalEvent(Object, Object)
+         */
+        public FlyweightVertexEvent(Object eventSource, VV vertex)
+        {
+            super(eventSource, vertex);
+        }
+
+        /**
+         * Sets the vertex of this event.
+         *
+         * @param vertex the vertex to be set.
+         */
+        protected void setVertex(VV vertex)
+        {
+            this.vertex = vertex;
+        }
+    }
+    
+    // -------------------------------------------------------------------------
+    /**
+     * Creates directed/undirected graph specifics according to the provided
+     * graph - directed/undirected, respectively. 
+     * 
+     * @param g the graph to create specifics for
+     *
+     * @return the created specifics
+     */
+    static <V, E> Specifics<V, E> createGraphSpecifics(Graph<V, E> g)
+    {
+        if (g instanceof DirectedGraph<?, ?>) {
+            return new DirectedSpecifics<>((DirectedGraph<V, E>) g);
+        } else {
+            return new UndirectedSpecifics<>(g);
+        }
+    }
+    
+    /**
+     * Provides unified interface for operations that are different in directed
+     * graphs and in undirected graphs.
+     */
+    abstract static class Specifics<VV, EE>
+    {
+        /**
+         * Returns the edges outgoing from the specified vertex in case of
+         * directed graph, and the edge touching the specified vertex in case of
+         * undirected graph.
+         *
+         * @param vertex the vertex whose outgoing edges are to be returned.
+         *
+         * @return the edges outgoing from the specified vertex in case of
+         * directed graph, and the edge touching the specified vertex in case of
+         * undirected graph.
+         */
+        public abstract Set<? extends EE> edgesOf(VV vertex);
+    }
+
+    /**
+     * An implementation of {@link Specifics} for a directed graph.
+     */
+    static class DirectedSpecifics<VV, EE>
+        extends Specifics<VV, EE>
+    {
+        private DirectedGraph<VV, EE> graph;
+
+        /**
+         * Creates a new DirectedSpecifics object.
+         *
+         * @param g the graph for which this specifics object to be created.
+         */
+        public DirectedSpecifics(DirectedGraph<VV, EE> g)
+        {
+            graph = g;
+        }
+
+        /**
+         * @see CrossComponentIterator.Specifics#edgesOf(Object)
+         */
+        @Override public Set<? extends EE> edgesOf(VV vertex)
+        {
+            return graph.outgoingEdgesOf(vertex);
+        }
+    }
+
+    /**
+     * An implementation of {@link Specifics} in which edge direction (if any)
+     * is ignored.
+     */
+    static class UndirectedSpecifics<VV, EE>
+        extends Specifics<VV, EE>
+    {
+        private Graph<VV, EE> graph;
+
+        /**
+         * Creates a new UndirectedSpecifics object.
+         *
+         * @param g the graph for which this specifics object to be created.
+         */
+        public UndirectedSpecifics(Graph<VV, EE> g)
+        {
+            graph = g;
+        }
+
+        /**
+         * @see CrossComponentIterator.Specifics#edgesOf(Object)
+         */
+        @Override public Set<EE> edgesOf(VV vertex)
+        {
+            return graph.edgesOf(vertex);
         }
     }
 }
